+++ src/lib/components/forms/asset-input/AssetInputNoOptionsMessage.tsx
@@ -2,9 +2,12 @@ import { Text } from "@chakra-ui/react";
 import type { NoticeProps } from "chakra-react-select";
 import { components } from "chakra-react-select";
 
-import type { AssetOption } from "lib/types";
+import type { SelectInputOption } from "../SelectInput";
+import type { AssetOptionValue } from "lib/types";
 
-export const AssetInputNoOptionsMessage = (props: NoticeProps<AssetOption>) => (
+export const AssetInputNoOptionsMessage = (
+  props: NoticeProps<SelectInputOption<AssetOptionValue>>
+) => (
   <components.NoOptionsMessage {...props}>
     <Text py={4}>No matching assets found</Text>
   </components.NoOptionsMessage>